use abstract class for default callback
