+++ demo-app/src/main/java/io/opentelemetry/android/demo/OtelDemoApplication.kt
@@ -59,10 +59,6 @@ class OtelDemoApplication : Application() {
         try {
             rum = otelRumBuilder.build()
             Log.d(TAG, "RUM session started: " + rum!!.rumSessionId)
-            val webSocketClient = WebSocketClient()
-            webSocketClient.connect()
-            webSocketClient.sendHello()
-            webSocketClient.sendHello()
         } catch (e: Exception) {
             Log.e(TAG, "Oh no!", e)
         }
