+++ instrumentation/view-click/src/main/kotlin/io/opentelemetry/android/instrumentation/view/click/WindowCallbackWrapper.kt
@@ -6,16 +6,19 @@
 package io.opentelemetry.android.instrumentation.view.click
 
 import android.os.Build.VERSION_CODES
-import android.view.MotionEvent
+import android.view.ActionMode
+import android.view.SearchEvent
 import android.view.Window.Callback
 import androidx.annotation.RequiresApi
 
 @RequiresApi(api = VERSION_CODES.M)
 class WindowCallbackWrapper(
     private val callback: Callback,
-) : Callback by callback {
-    override fun dispatchTouchEvent(event: MotionEvent?): Boolean {
-        ViewClickEventGenerator.generateClick(event)
-        return callback.dispatchTouchEvent(event)
-    }
+) : DefaultWindowCallback(callback) {
+    override fun onSearchRequested(searchEvent: SearchEvent?): Boolean = callback.onSearchRequested(searchEvent)
+
+    override fun onWindowStartingActionMode(
+        callback: ActionMode.Callback?,
+        type: Int,
+    ): ActionMode? = this.callback.onWindowStartingActionMode(callback, type)
 }
